backend/eval/requirements.txt
openai
supabase
ragas
datasets
python-dotenv
langchain-openai
